requirements.txt
openai==0.27.2
discord.py==2.2.2
